+++ GridboxConnectorAddon/GridboxConnector/ha_viessmann_gridbox_connector.py
@@ -13,8 +13,9 @@ class HAViessmannGridboxConnector:
     consumption_household_sensor: Sensor
     total_consumption_household_sensor: Sensor
     direct_consumption_household_sensor: Sensor
-    direct_consumtion_heatpump_sensor: Sensor
-    direct_consumtion_rate_sensor: Sensor
+    direct_consumption_heatpump_sensor: Sensor
+    direct_consumption_ev_sensor: Sensor
+    direct_consumption_rate_sensor: Sensor
     self_supply_sensor: Sensor
     self_consumtion_rate_sensor: Sensor
     self_sufficiency_rate_sensor: Sensor
@@ -28,60 +29,43 @@ class HAViessmannGridboxConnector:
 
         production_sensor_info = SensorInfo(name="Production", device_class="power",
                                             unique_id="gridbox_production", device=self.device_info, unit_of_measurement="W")
-        production_settings = Settings(
-            mqtt=mqtt_settings, entity=production_sensor_info)
+        production_settings = Settings(mqtt=mqtt_settings, entity=production_sensor_info)
 
-        grid_sensor_info = SensorInfo(name="Grid", device_class="power",
-                                      unique_id="gridbox_grid", device=self.device_info, unit_of_measurement="W")
+        grid_sensor_info = SensorInfo(name="Grid", device_class="power", unique_id="gridbox_grid", device=self.device_info, unit_of_measurement="W")
         grid_settings = Settings(mqtt=mqtt_settings, entity=grid_sensor_info)
 
-        photovoltaic_sensor_info = SensorInfo(name="Photovoltaic", device_class="power",
-                                              unique_id="gridbox_photovoltaic", device=self.device_info, unit_of_measurement="W")
-        photovoltaic_settings = Settings(
-            mqtt=mqtt_settings, entity=photovoltaic_sensor_info)
+        photovoltaic_sensor_info = SensorInfo(name="Photovoltaic", device_class="power", unique_id="gridbox_photovoltaic", device=self.device_info, unit_of_measurement="W")
+        photovoltaic_settings = Settings(mqtt=mqtt_settings, entity=photovoltaic_sensor_info)
 
         # consumption
-        consumption_household_sensor_info = SensorInfo(
-            name="Consumption", device_class="power", unique_id="gridbox_consumption_household", device=self.device_info, unit_of_measurement="W")
-        consumption_household_settings = Settings(
-            mqtt=mqtt_settings, entity=consumption_household_sensor_info)
+        consumption_household_sensor_info = SensorInfo(name="Consumption", device_class="power", unique_id="gridbox_consumption_household", device=self.device_info, unit_of_measurement="W")
+        consumption_household_settings = Settings(mqtt=mqtt_settings, entity=consumption_household_sensor_info)
 
-        total_consumption_household_sensor_info = SensorInfo(
-            name="Total Consumption", device_class="power", unique_id="total_consumption_household", device=self.device_info, unit_of_measurement="W")
-        total_consumption_household_settings = Settings(
-            mqtt=mqtt_settings, entity=total_consumption_household_sensor_info)
+        total_consumption_household_sensor_info = SensorInfo(name="Total Consumption", device_class="power", unique_id="total_consumption_household", device=self.device_info, unit_of_measurement="W")
+        total_consumption_household_settings = Settings(mqtt=mqtt_settings, entity=total_consumption_household_sensor_info)
 
         # Direct Consumption
-        direct_consumption_household_sensor_info = SensorInfo(
-            name="DirectConsumptionHousehold", device_class="power", unique_id="gridbox_direct_consumption_household", device=self.device_info, unit_of_measurement="W")
-        direct_consumption_household_settings = Settings(
-            mqtt=mqtt_settings, entity=direct_consumption_household_sensor_info)
+        direct_consumption_household_sensor_info = SensorInfo(name="DirectConsumptionHousehold", device_class="power", unique_id="gridbox_direct_consumption_household", device=self.device_info, unit_of_measurement="W")
+        direct_consumption_household_settings = Settings(mqtt=mqtt_settings, entity=direct_consumption_household_sensor_info)
 
-        direct_consumption_heatpump_sensor_info = SensorInfo(
-            name="DirectConsumptionHeatPump", device_class="power", unique_id="gridbox_direct_consumption_heatpump", device=self.device_info, unit_of_measurement="W")
-        direct_consumption_heatpump_settings = Settings(
-            mqtt=mqtt_settings, entity=direct_consumption_heatpump_sensor_info)
+        direct_consumption_heatpump_sensor_info = SensorInfo(name="DirectConsumptionHeatPump", device_class="power", unique_id="gridbox_direct_consumption_heatpump", device=self.device_info, unit_of_measurement="W")
+        direct_consumption_heatpump_settings = Settings(mqtt=mqtt_settings, entity=direct_consumption_heatpump_sensor_info)
 
-        direct_consumption_rate_sensor_info = SensorInfo(
-            name="DirectConsumptionRate", device_class="power_factor", unique_id="gridbox_direct_consumption_rate", device=self.device_info, unit_of_measurement="%")
-        direct_consumption_rate_settings = Settings(
-            mqtt=mqtt_settings, entity=direct_consumption_rate_sensor_info)
+        direct_consumption_ev_sensor_info = SensorInfo(name="DirectConsumptionEV", device_class="power", unique_id="gridbox_direct_consumption_ev", device=self.device_info, unit_of_measurement="W")
+        direct_consumption_ev_settings = Settings(mqtt=mqtt_settings, entity=direct_consumption_ev_sensor_info)
+
+        direct_consumption_rate_sensor_info = SensorInfo(name="DirectConsumptionRate", device_class="power_factor", unique_id="gridbox_direct_consumption_rate", device=self.device_info, unit_of_measurement="%")
+        direct_consumption_rate_settings = Settings(mqtt=mqtt_settings, entity=direct_consumption_rate_sensor_info)
 
         # Self Consumption
-        self_supply_sensor_info = SensorInfo(name="SelfSupply", device_class="power",
-                                             unique_id="gridbox_self_supply", device=self.device_info, unit_of_measurement="W")
-        self_supply_settings = Settings(
-            mqtt=mqtt_settings, entity=self_supply_sensor_info)
+        self_supply_sensor_info = SensorInfo(name="SelfSupply", device_class="power",unique_id="gridbox_self_supply", device=self.device_info, unit_of_measurement="W")
+        self_supply_settings = Settings(mqtt=mqtt_settings, entity=self_supply_sensor_info)
 
-        self_consumption_rate_sensor_info = SensorInfo(
-            name="SelfConsumptionRate", device_class="power_factor", unique_id="gridbox_self_consumption_rate", device=self.device_info, unit_of_measurement="%")
-        self_consumption_rate_settings = Settings(
-            mqtt=mqtt_settings, entity=self_consumption_rate_sensor_info)
+        self_consumption_rate_sensor_info = SensorInfo(name="SelfConsumptionRate", device_class="power_factor", unique_id="gridbox_self_consumption_rate", device=self.device_info, unit_of_measurement="%")
+        self_consumption_rate_settings = Settings(mqtt=mqtt_settings, entity=self_consumption_rate_sensor_info)
 
-        self_sufficiency_rate_sensor_info = SensorInfo(
-            name="SelfSufficiencyRate", device_class="power_factor", unique_id="gridbox_self_sufficiency_rate", device=self.device_info, unit_of_measurement="%")
-        self_sufficiency_rate_settings = Settings(
-            mqtt=mqtt_settings, entity=self_sufficiency_rate_sensor_info)
+        self_sufficiency_rate_sensor_info = SensorInfo(name="SelfSufficiencyRate", device_class="power_factor", unique_id="gridbox_self_sufficiency_rate", device=self.device_info, unit_of_measurement="%")
+        self_sufficiency_rate_settings = Settings(mqtt=mqtt_settings, entity=self_sufficiency_rate_sensor_info)
 
         # Instantiate the sensors
 
@@ -94,22 +78,16 @@ class HAViessmannGridboxConnector:
             mqtt_settings, self.device_info, "sum", "")
 
         # Consumption
-        self.consumption_household_sensor = Sensor(
-            consumption_household_settings)
-        self.total_consumption_household_sensor = Sensor(
-            total_consumption_household_settings)
-        self.direct_consumption_household_sensor = Sensor(
-            direct_consumption_household_settings)
-        self.direct_consumtion_heatpump_sensor = Sensor(
-            direct_consumption_heatpump_settings)
-        self.direct_consumtion_rate_sensor = Sensor(
-            direct_consumption_rate_settings)
+        self.consumption_household_sensor = Sensor(consumption_household_settings)
+        self.total_consumption_household_sensor = Sensor(total_consumption_household_settings)
+        self.direct_consumption_household_sensor = Sensor(direct_consumption_household_settings)
+        self.direct_consumption_heatpump_sensor = Sensor(direct_consumption_heatpump_settings)
+        self.direct_consumption_ev_sensor = Sensor(direct_consumption_ev_settings)
+        self.direct_consumption_rate_sensor = Sensor(direct_consumption_rate_settings)
 
         self.self_supply_sensor = Sensor(self_supply_settings)
-        self.self_consumtion_rate_sensor = Sensor(
-            self_consumption_rate_settings)
-        self.self_sufficiency_rate_sensor = Sensor(
-            self_sufficiency_rate_settings)
+        self.self_consumtion_rate_sensor = Sensor(self_consumption_rate_settings)
+        self.self_sufficiency_rate_sensor = Sensor(self_sufficiency_rate_settings)
 
     def update_sensors(self, measurement: dict):
         if "production" in measurement:
@@ -125,9 +103,11 @@ class HAViessmannGridboxConnector:
         if "directConsumptionHousehold" in measurement:
             self.direct_consumption_household_sensor.set_state(float(measurement.get("directConsumptionHousehold", "0")))
         if "directConsumptionHeatPump" in measurement:
-            self.direct_consumtion_heatpump_sensor.set_state(float(measurement.get("directConsumptionHeatPump", "0")))
+            self.direct_consumption_heatpump_sensor.set_state(float(measurement.get("directConsumptionHeatPump", "0")))
+        if "directConsumptionEV" in measurement:
+            self.direct_consumption_ev_sensor.set_state(float(measurement.get("directConsumptionEV", "0")))
         if "directConsumptionRate" in measurement:
-            self.direct_consumtion_rate_sensor.set_state(float(measurement.get("directConsumptionRate", "0"))*100)
+            self.direct_consumption_rate_sensor.set_state(float(measurement.get("directConsumptionRate", "0"))*100)
 
         if "selfSupply" in measurement:
             self.self_supply_sensor.set_state(float(measurement.get("selfSupply", "")))
@@ -143,6 +123,7 @@ class HAViessmannGridboxConnector:
             power = float(battery.get("power", "0"))
             remaining_charge = float(battery.get("remainingCharge", "0"))
             self.battery_sum.set_states(state_of_charge, capacity, power, remaining_charge)
+            
         if "batteries" in measurement:
             batteries: list = measurement.get("batteries", [])
             for index, battery in enumerate(batteries):